Conflicts resolved in 5 files:
- schema.sql: kept main's SMALLINT caffeine, added social + knowledge columns
- engine.py: kept our 3-field decay (caffeine -2, social -1, knowledge -1)
- shop.py: accepted main's fund endpoint + improved transfer, kept our
  coffee→caffeine and pizza→hamburger changes
- api.js: accepted main's fundDev + transferNxt (with token_id params)
- MyDevs.jsx: accepted main's version (has fund modal, transfer modal,
  wallet balance helpers built-in, hamburger already renamed)

+++ backend/db/schema.sql
@@ -157,10 +157,10 @@ CREATE TABLE devs (
     last_message        TEXT,
     last_message_channel chat_channel_enum,
 
-    -- Vitals (decayed by engine each salary tick)
-    caffeine            INTEGER NOT NULL DEFAULT 50,
-    social              INTEGER NOT NULL DEFAULT 50,
-    knowledge           INTEGER NOT NULL DEFAULT 50,
+    -- Vital stats (decay over time)
+    caffeine            SMALLINT NOT NULL DEFAULT 50,
+    social              SMALLINT NOT NULL DEFAULT 50,
+    knowledge           SMALLINT NOT NULL DEFAULT 50,
 
     -- Game mechanics
     pc_health           SMALLINT NOT NULL DEFAULT 100,
